GUI/requirements.txt
pyserial
tkinter 